@@ -3,37 +3,40 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
-    "@hookform/resolvers": "^2.9.8",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/joy": "^5.0.0-alpha.47",
-    "@mui/lab": "^5.0.0-alpha.72",
-    "@mui/material": "^5.2.7",
-    "@reduxjs/toolkit": "^1.8.1",
-    "@stripe/react-stripe-js": "^1.13.0",
-    "@stripe/stripe-js": "^1.42.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.3",
-    "@types/node": "^16.11.17",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@hookform/resolvers": "^3.0.0",
+    "@mui/icons-material": "^5.11.11",
+    "@mui/joy": "^5.0.0-alpha.73",
+    "@mui/lab": "^5.0.0-alpha.124",
+    "@mui/material": "^5.11.15",
+    "@reduxjs/toolkit": "^1.9.3",
+    "@stripe/react-stripe-js": "^2.1.0",
+    "@stripe/stripe-js": "^1.52.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.32",
+    "@types/react-dom": "^18.0.11",
     "@types/react-infinite-scroller": "^1.2.3",
-    "@types/react-router-dom": "^5.3.2",
-    "axios": "^0.21.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.34.2",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-slick": "^0.23.10",
+    "axios": "^1.3.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.9",
     "react-infinite-scroller": "^1.2.6",
-    "react-redux": "^7.2.8",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "5.0.0",
-    "react-toastify": "^8.1.0",
-    "typescript": "^4.5.4",
-    "web-vitals": "^2.1.2",
-    "yup": "^0.32.11"
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "^5.0.1",
+    "react-slick": "^0.29.0",
+    "react-toastify": "^9.1.2",
+    "slick-carousel": "^1.8.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^3.3.0",
+    "yup": "^1.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
