@@ -1,28 +1,29 @@
+import ReactDOM from 'react-dom/client';
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './app/layout/styles.css';
 import './app/models/product';
-import App from './app/layout/App';
 import reportWebVitals from './reportWebVitals';
-import { Router } from 'react-router-dom';
-import { createBrowserHistory } from "history";
+import { RouterProvider } from 'react-router-dom';
 import { StoreContextProvider } from './app/context/StoreContext';
 import { Provider } from 'react-redux';
 import { store } from './app/REDUX/configureStore';
+import "slick-carousel/slick/slick.css";
+import "slick-carousel/slick/slick-theme.css";
+import { router } from './Routes/Routes';
 
-export const history = createBrowserHistory();
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Router history={history}>
-      <StoreContextProvider>
-        <Provider store={store}>
-        <App />
-        </Provider>
-      </StoreContextProvider>
-    </Router>
-  </React.StrictMode>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+
+root.render(
+<React.StrictMode>
+ <StoreContextProvider>
+  <Provider store = {store} > {/* redux provider */}
+    <RouterProvider router = {router} />
+  </Provider>
+ </StoreContextProvider>
+ </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
