@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.IO;
 using System.Text;
 using API.Data;
 using API.Entities;
@@ -8,11 +7,9 @@ using API.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
@@ -106,6 +103,17 @@ namespace API
                 app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "WebAPIv5 v1"));
             }
 
+            //for using formFiles
+            // app.UseStaticFiles(new StaticFileOptions(){
+            //     FileProvider = new PhysicalFileProvider
+            //     (Path.Combine(Directory.GetCurrentDirectory(), @"Uploads")), 
+            //     RequestPath = new PathString("/Uploads")
+            // });
+           // 
+
+           app.UseDefaultFiles(); // telling (ASP.NETCore) API-Server to use default files.
+           app.UseStaticFiles(); // telling (ASP.NETCore) API-Server to use static files.
+
            // app.UseHttpsRedirection();
             //app.UseResponseCompression();
             app.UseRouting();
@@ -115,22 +123,17 @@ namespace API
             opt.AllowAnyHeader().AllowAnyMethod().AllowCredentials().WithOrigins("http://localhost:3000");
             });
 
-            //for using formFiles
-            app.UseStaticFiles(new StaticFileOptions(){
-                FileProvider = new PhysicalFileProvider
-                (Path.Combine(Directory.GetCurrentDirectory(), @"Uploads")), 
-                RequestPath = new PathString("/Uploads")
-            });
-           // 
             
             app.UseAuthentication();
 
             app.UseAuthorization();
 
             app.UseEndpoints(endpoints =>
-            {
+            { 
                 endpoints.MapControllers();
+                endpoints.MapFallbackToController("Index", "Fallback");
             });
+
         }
     }
 }
