@@ -2,13 +2,13 @@ import axios, { AxiosError, AxiosResponse } from "axios";
 import { FieldValues } from "react-hook-form";
 import { toast } from "react-toastify";
 import { URLSearchParams } from "url";
-import { history } from "../..";
 import { PaginatedResponse } from "../models/pagination";
 import { store } from "../REDUX/configureStore";
+import { router } from "../../Routes/Routes";
 
 const sleep =()=> new Promise(resolve => setTimeout(resolve, 500));
 
-axios.defaults.baseURL='http://localhost:5000/api/';
+axios.defaults.baseURL = process.env.REACT_APP_API_BASE_URL;
 axios.defaults.withCredentials =true;
 
 const responseBody = (response : AxiosResponse) => response.data;
@@ -65,7 +65,7 @@ axios.interceptors.request.use(config => {
 });
 
 axios.interceptors.response.use(async response =>{
-    await sleep();
+    if (process.env.NODE_ENV === 'development') await sleep();
     const pagination = response.headers['pagination']; //axios always accept the lower case header namings
     if (pagination) {
         response.data = new PaginatedResponse(response.data, JSON.parse(pagination));
@@ -75,7 +75,7 @@ axios.interceptors.response.use(async response =>{
     return response;
 }, (error: AxiosError)=>{
      
-    const {data, status} = error.response!;
+    const {data, status} : any = error.response!;
     switch(status){
 
         case 400:
@@ -101,10 +101,7 @@ axios.interceptors.response.use(async response =>{
         break;
 
         case 500:
-            history.push({
-                pathname : '/server-error',
-                state : {error: data},  
-            });
+            router.navigate('/server-error', {state : {error: data}});
         break;
 
         case 404: toast.error(data.title, {
