@@ -2,38 +2,20 @@ import { ThemeProvider, createTheme } from "@mui/material/styles";
 import { Container, CssBaseline } from "@mui/material";
 import Header from "./Header";
 import { useCallback, useEffect, useState } from "react";
-import AboutPage from "../../features/about/AboutPage";
-import { Route } from "react-router-dom";
-import ProductDetails from "../../features/catalog/ProductDetails";
-import ContactPage from "../../features/Contact/ContactPage";
-import CollectionPage from "../../features/Collections/CollectionPage";
-import HomePage from "../../features/home/HomePage";
-import Catalog from "../../features/catalog/Catalog";
+import { Outlet, Route, useLocation } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 import 'react-toastify/dist/ReactToastify.css';
-import ServerError from "../error/ServerError";
-import NotFound from "../error/NotFound";
-import { Switch } from "react-router-dom";
-import BasketPage from "../../features/BasketPage/BasketPage";
 import Loading from "./Loading";
 import { fetchBasketAsync } from "../../features/BasketPage/BasketSlice";
 import { useAppDispatch } from "../REDUX/configureStore";
-import Login from "../../features/account/Login/Login";
-import Register from "../../features/account/Register/Register";
 import { fetchUserAsync } from "../../features/account/accountSlice";
-import PrivateRoute from "../../features/account/PrivateRoute";
-import OrdersPage from "../../features/Order/OrdersPage";
-import OrderDetails from "../../features/Order/OrderDetails";
-import CheckoutWrapper from "../../features/CheckoutPage/CheckoutWrapper";
+import HomePage from "../../features/home/HomePage";
 
-
-
-
-
- function App() {
+function App() {
  
 const [initializing, setinitializing] = useState(true);
-/*const { setBasket } = useStoreContext();*/
+
+const location = useLocation(); // using react-router-dom
 const dispatch = useAppDispatch();
 const [darkMode , setDarkMode]= useState(false);
 
@@ -48,11 +30,6 @@ const InitializeApplication = useCallback(async () => {
 useEffect(()=>{
 InitializeApplication().then(() => setinitializing(false));
 }, [InitializeApplication])
-
-if (initializing) return <Loading message="Staging App ..." />
-
-
-
 
   const paletteType= darkMode ? 'dark' : 'light';
   const theme= createTheme({
@@ -74,24 +51,11 @@ if (initializing) return <Loading message="Staging App ..." />
       <ToastContainer position ='bottom-right' hideProgressBar />
     <CssBaseline/>
      <Header darkMode={darkMode} HandleThemeChange={HandleThemeChange} />
-     <Container>
-     <Switch>
-     <Route exact path='/' component={HomePage} />
-     <Route exact path='/catalog' component= {Catalog} />
-     <Route exact path='/catalog/:id' component={ProductDetails} />
-     <Route exact path='/about' component={AboutPage} />
-     <Route exact path='/Contact' component={ContactPage} />
-     <Route exact path='/Collections' component={CollectionPage} />
-     <Route exact path ='/server-error' component={ServerError} />
-     <Route exact path ='/BasketPage' component={BasketPage} />
-     <Route exact path ='/account/Login' component={Login} />
-     <Route exact path ='/account/Register' component={Register} />
-     <PrivateRoute exact path='/CheckoutPage' component={CheckoutWrapper} />
-     <PrivateRoute exact path='/Order' component={OrdersPage} />
-     <PrivateRoute exact path='/Order/:id' component={OrderDetails} />
-     <Route component ={NotFound} />
-     </Switch>
-     </Container>
+     {initializing ? <Loading message="Staging App ..." />
+                   : location.pathname === '/' ?  <HomePage />
+                   : ( <Container>
+                         <Outlet />
+                       </Container> )}
      </ThemeProvider>
   );
 }
